service/requirements.txt
psycopg2-binary
python-dotenv
confluent-kafka
redis
